requirements.txt
anthropic>=0.25.0
python-dotenv~=1.1.1
asyncio
mcp
